refactor: 추천 여행지 서비스 함수 fetchRecommendationList, fetchRecommendationDetail 분리
